+++ docs/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 * :feature:`3254` Users can now add a custom Covalent API key in rotki.
 * :feature:`3416` Users will now be able to import ShapeShift trade history CSVs exported from shapeshift.com.
 * :feature:`1345` Users will now be able to import uphold transaction history CSVs exported from uphold.com.
+* :feature:`-` Users will now be able to track balances locked in pickle's DILL.
 * :bug:`3239` Users will now see only the relevant addresses suggested when filtering DeFi deposits.
 * :bug:`3340` Users will now be able to properly see the favicon when accessing the frontend using Firefox.
 * :bug:`3266` Users will now be able to properly access most of the top bar content using a mobile device.